Issue-51: Prints Process Logs in RuntimeException Message
This resolves the issue: https://github.com/kstyrc/embedded-redis/issues/51

While resolving logs at the process level is preferred as specified in the pull-request: https://github.com/kstyrc/embedded-redis/pull/113 - this fix takes a much more conservative approach and prints the redis process logs only in case of an error/exception when the server startup pattern is not found. That way we do not need to worry about logs in case of successful redis server starts.

Dist updated to trusty to resolves travis jdk8 build issue caused by xenial dist

Removed sudo and curl commands as they are not longer supported.
jpm4j is not longer functional and sudo is deprecated in travis